[ CHORE ] example management #18
