+++ examples/real-time-communication/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>RTC :: Boardy Example</title>
+  <style>
+    body {margin:0;padding:0;}
+    canvas {margin:20px;background-color:#f2f2f2;}
+  </style>
+</head>
+<body>
+  <canvas id="leader-canvas" width="600" height="600"></canvas>
+  <canvas id="member-canvas-1" width="200" height="200"></canvas>
+  <canvas id="member-canvas-2" width="300" height="300"></canvas>
+</body>
+</html>
