+++ examples/simple/index.html
@@ -3,16 +3,10 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Boardy Example</title>
+  <title>Simple :: Boardy Example</title>
   <style>
-    body {
-      margin: 0;
-      padding: 0;
-    }
-    canvas {
-      margin: 20px;
-      background-color: #f2f2f2;
-    }
+    body {margin:0;padding:0;}
+    canvas {margin:20px;background-color:#f2f2f2;}
   </style>
 </head>
 <body>
