+++ src/modules/Rasterizer/Rasterizer.ts
@@ -8,6 +8,16 @@ export default class Rasterizer {
   }
 
   public rasterize() {
-    this.context.ctx.screen.drawImage(this.context.canvas.offscreen, 0, 0);
+    this.context.ctx.screen.drawImage(
+      this.context.canvas.offscreen,
+      0,
+      0,
+      this.context.resolution[0],
+      this.context.resolution[1],
+      0,
+      0,
+      this.context.size[0],
+      this.context.size[1],
+    );
   }
 }
