+++ tsconfig.json
@@ -19,7 +19,7 @@
     ]
   },
   "include": [
-      "src/**/*"
+      "src/**/*", "examples/**/*"
   ],
   "exclude": [
       "node_modules",
